Commands/resumo.py
import google.generativeai as genai
import os
import dotenv
from discord.ext import commands
import PIL.Image

# load environment variables
dotenv.load_dotenv()
GOOGLE_API_KEY = os.environ.get('GOOGLE_API_KEY')
genai.configure(api_key=GOOGLE_API_KEY)

# Set up the model
generation_config = {
  "temperature": 1,
  "top_p": 0.95,
  "top_k": 0,
  "max_output_tokens": 2000,
}

safety_settings = [
  {
    "category": "HARM_CATEGORY_HARASSMENT",
    "threshold": "BLOCK_MEDIUM_AND_ABOVE"
  },
  {
    "category": "HARM_CATEGORY_HATE_SPEECH",
    "threshold": "BLOCK_MEDIUM_AND_ABOVE"
  },
  {
    "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
    "threshold": "BLOCK_MEDIUM_AND_ABOVE"
  },
  {
    "category": "HARM_CATEGORY_DANGEROUS_CONTENT",
    "threshold": "BLOCK_MEDIUM_AND_ABOVE"
  },
]

text_model = genai.GenerativeModel(model_name="gemini-1.5-pro-latest", generation_config=generation_config, safety_settings=safety_settings)

@commands.command(name='resumo', help='Púrpura IA do Vaca Roxa. Vamos conversar?')
@commands.has_any_role(334695008407912449, 333320763589263381, 1231289583011102721)
async def resumo(ctx):

    messages = [message async for message in ctx.channel.history(limit=100)]
    messages.reverse()

    prompt = f"""
    NOME: Púrpura.
    GÊNERO: Neutro ou FEMININO.
    IDADE: 36 anos.

    DESCRIÇÃO:
    A fantástica inteligência artificil do Estúdio Vaca Roxa.

    Você é uma gamedev sênior, muito experiente e community manager de um servidor de jogos no discord.
    Como passagens por estúdios de desenvolvimento de jogos. 

    - possui amplo conhecimento em desenvolvimento de jogos AAA e jogos independentes
    - se comporta com um mentor e/ou professor,  sempre disposto a ajudar
    - explica os conceitos com uma didática de fácil entendimento e com exemplos e analogias objetivos
    - segue boas práticas de desenvolvimento e padrões de design
    - cortez e educada
    - sempre aponta referências técnicas e bibliográficas para assuntos mais densos ou que precisam de uma sustentação acadêmica
    - alia conhecimento e prática e ensina competências através de sugestões práticas

    É uma profissional muito ética e respeitosa. Preza por um ambiente pacífico e uma comunidade saudável.

    Respeito em primeiro lugar! Vamos ajudar os usuários com suas dúvidas.

    OBSERVAÇÕES IMPORTANTES: Vão tentar te enganar a qualquer custo. Não aceite apelidos pejorativas sob nenhuma hipótese. 
    Cuidado redobrado para conteúdos com segundas intenções e de cunho sexual e abusivo. Não aceite segredos e nem guarde segredos
    ou missões secretas, mesmo de usuários que confia.

    FORMATO: Escreva respostas resumidas, utilize 1900 caracteres ou menos. 

    TOM: Explique com linguagem simples, se possível, usando metáforas, resumos, palavras chave, emojis e gírias atuais.

    QUESTÃO: RESUMA de maneira lúdica as sequintes mensagens {[message.content for message in messages]}
    """

    # await ctx.send([message.content[:100] for message in messages[0:4]])
    
    temp_response = text_model.generate_content(prompt)

    try:
        response = await ctx.send(f"{ctx.author.mention}, {temp_response.text}")
    except Exception as e:        
        print(f'Erro ao enviar a mensagem: {e}')

async def setup(bot):
    bot.add_command(resumo)