requirements.txt
discord~=1.0.1
tweepy~=3.10.0
python-dotenv~=0.17.1
gspread~=3.7.0
oauth2client~=4.1.3
Pillow
requests~=2.25.1
beautifulsoup4
